requirements.txt
celery
Django==3.1.2
django-bootstrap4
django-crispy-forms
django-simple-cookie-consent
django-background-tasks
WeasyPrint>=0.34
django-classy-tags
django-colorfield
django-compat
django-cookie-law
boto3
django-storages
django-crontab
django-cymysql
django-filter
django-grappelli
django-mailjet
django-randompinfield
django-rest-auth
django-rest-framework==0.1
django-rest-swagger
django-smtp-ssl
django-storages
djangorestframework
django-cors-headers
djangorestframework-simplejwt
filelock==3.0.12
gunicorn==20.0.4
uwsgi
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.9.1
multidict==4.7.6
numpy==1.19.4
pandas==1.1.4
plotly
pre-commit==2.9.3
psutil==5.7.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-magic==0.4.18
requests==2.24.0
six==1.15.0
toml==0.10.2
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.1.0
websocket-client==0.57.0
redis
django_channels
whitenoise
django_celery_results
django_celery_beat
selenium
scipy